Revert "Switched to directxk color"
This reverts commit 5f5141683adfe4bad587b088dae9b19d974e61b6.

+++ src/Header Files/HairDrawing/BrushData.h
@@ -13,7 +13,7 @@ enum class BrushMode
 
 struct BrushData
 {
-	float strength;
+	int strength;
 	float radius;
 	PaintChannel paintChannel;
 	BrushMode mode;
